requirements.txt
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.13.2
scikit-learn==1.3.0
opencv-python==4.8.1.78
gradio==3.41.2
Pillow==10.0.1
protobuf==4.24.3
h5py==3.10.0
